refactor(vnext): relocate config to src/vnext/config.ts (#623 step 3)
Pure relocate, no logic change. Decided against the per-domain split
(agent/binding/sandbox/...) — at 677 LOC the file is small enough that
splitting now would be premature. openclaw's 140-file config/ tree is the
result of years of accretion, not initial design. Defer splitting until a
specific domain actually grows.

- src/core/config.{ts,test.ts} → src/vnext/
- types.ts dependency stays in src/core/ (later step)
- update 7 importers + 4 core-internal imports
- 59/59 tests pass
